style: more kotlin style
