+++ src/main/kotlin/me/rerere/unocssintellij/highlighting/UnocssAttributeExternalAnnotator.kt
@@ -10,18 +10,14 @@ import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.TextRange
 import com.intellij.openapi.util.text.StringUtil
 import com.intellij.psi.PsiDocumentManager
-import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
-import com.intellij.psi.css.impl.CssElementTypes
-import com.intellij.psi.util.elementType
-import com.intellij.psi.xml.XmlElementType
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.withTimeoutOrNull
 import me.rerere.unocssintellij.UnocssService
 import me.rerere.unocssintellij.settings.UnocssSettingsState
 import me.rerere.unocssintellij.util.MatchedPosition
+import me.rerere.unocssintellij.util.isUnocssCandidate
 import me.rerere.unocssintellij.util.getMatchedPositions
-import me.rerere.unocssintellij.util.isJsString
 
 data class UnocssInitInfo(
     val project: Project,
@@ -78,26 +74,13 @@ class UnocssAttributeExternalAnnotator : ExternalAnnotator<UnocssInitInfo, Unocs
     }
 
     override fun apply(file: PsiFile, annotationResult: UnocssAnnotationResult?, holder: AnnotationHolder) {
-        annotationResult?.annotations?.forEach {
-            val element = file.findElementAt(it.start) ?: return@forEach
-            if (isValidElement(element)) {
+        annotationResult?.annotations
+            ?.filter { file.findElementAt(it.start)?.isUnocssCandidate() ?: false }
+            ?.forEach {
                 holder.newSilentAnnotation(HighlightSeverity.TEXT_ATTRIBUTES)
                     .range(TextRange.create(it.start, it.end))
                     .textAttributes(DefaultLanguageHighlighterColors.HIGHLIGHTED_REFERENCE)
                     .create()
             }
-        }
     }
-}
-
-private val VALID_ELEMENT_TYPES = setOf(
-    XmlElementType.XML_ATTRIBUTE_VALUE_TOKEN,
-    XmlElementType.XML_NAME,
-    CssElementTypes.CSS_STRING_TOKEN,
-    CssElementTypes.CSS_IDENT,
-)
-
-private fun isValidElement(element: PsiElement): Boolean {
-    val elementType = element.elementType
-    return elementType in VALID_ELEMENT_TYPES || isJsString(element)
 }
