+++ src/main/kotlin/me/rerere/unocssintellij/references/UnocssThemeBreakpointsConfigReference.kt
@@ -4,13 +4,16 @@ import com.intellij.codeInsight.lookup.LookupElement
 import com.intellij.codeInsight.lookup.LookupElementBuilder
 import com.intellij.lang.javascript.psi.JSObjectLiteralExpression
 import com.intellij.openapi.util.TextRange
-import com.intellij.psi.*
+import com.intellij.psi.PsiElement
+import com.intellij.psi.PsiElementResolveResult
+import com.intellij.psi.ResolveResult
+import me.rerere.unocssintellij.util.UnoConfigHelper
 
-class UnocssThemeBreakpointsConfigReference(element: PsiElement, textRange: TextRange)
-    : UnocssThemeConfigReference(element, textRange) {
+class UnocssThemeBreakpointsConfigReference(element: PsiElement, textRange: TextRange) :
+    UnocssThemeConfigReference(element, textRange) {
 
     override fun multiResolve(incompleteCode: Boolean): Array<ResolveResult> {
-        val themeConfigValue = UnoConfigPsiHelper.findThemeConfig(element) ?: return emptyArray()
+        val themeConfigValue = UnoConfigHelper.findThemeConfig(element) ?: return emptyArray()
 
         val results = mutableListOf<ResolveResult>()
         if (themeConfigValue is JSObjectLiteralExpression) {
@@ -31,7 +34,7 @@ class UnocssThemeBreakpointsConfigReference(element: PsiElement, textRange: Text
     }
 
     override fun getVariants(): Array<Any> {
-        val themeConfigValue = UnoConfigPsiHelper.findThemeConfig(element) ?: return emptyArray()
+        val themeConfigValue = UnoConfigHelper.findThemeConfig(element) ?: return emptyArray()
         if (themeConfigValue !is JSObjectLiteralExpression) return emptyArray()
 
         val variants = mutableListOf<LookupElement>()
